Backend/feat/record dto fix

See merge request s11-final/S11P31D102!63

+++ backend/runIt/src/main/java/com/ssafy/runit/domain/pace/dto/response/PaceResponse.java
@@ -6,10 +6,10 @@ import java.util.List;
 
 @Builder
 public record PaceResponse(
-    Double bpmList,
-    Double durationList
+    Integer bpmList,
+    Integer durationList
 ) {
-    public static PaceResponse isEntity(Double bpmLs, Double durationLs){
+    public static PaceResponse isEntity(Integer bpmLs, Integer durationLs){
         return PaceResponse.builder()
                 .bpmList(bpmLs)
                 .durationList(durationLs)
